ISSUE-285: ConfigureAppLocale Domain command
