+++ src/Domain/App/ConfigureAppLocale/ConfigureAppLocale.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\App\ConfigureAppLocale;
+
+use App\Infrastructure\CQRS\Bus\DomainCommand;
+
+final class ConfigureAppLocale extends DomainCommand
+{
+}
